refactor: player testFixture 분리
